Created FakeHttpResponse
Because a HttpResponse is the best "abstraction" we can work with when
handling Server Sent Events, we have to be able to fake it for testing
the server.

Sadly, when using HttpResponse, we are coupled to ASP.NET Core. But I
think this one is used in 99% of the use cases anyway.