Moved serialization of events to own classes
There are now three classes Event, Comment, and WaitRequest. All three
know how they should be written to a stream. Therefore, we can use
polymorphism.

I wanted to achieve that EventSerializer does not have to handle
serialization of all events. The change is also in favor of the
Open-Closed-Principle. The problems are that the IEvent classes now have
an additional responsibility and I added more complexity.