Removed HttpContext handling from Server
The Client class now handles all the HttpContext related stuff. The
Server only takes a Client and sends Events to its stream.

The user of the Server class needs to manage the clients. Later we will
provide an HttpClientManager that keeps track of different clients and
handles things like aborted connections.