Set LastEventId property if present in request
The HttpRequest can contain a Last-Event-Id header. If it is present, we
set the associated property of the HttpClient object.

The user of the HttpClient can do some actions if the property is set,
e.g. send all the events that happened after the event with id 'LastEventId'.

But this is none of the HttpClient's or Server's business.