Use clients of ClientManager when Broadcasting
An interaction now looks like this:
var et = new EventTransmitter();
et.Clients.Add(new Client());
et.BroadcastEvent("Hello Worlds!");

The advantage of this approach is that the user does not have to deal
with client management, for example watching if the connection is
closed.

I only hope that this API is intuitive.