Handle special case when only carriage return is sent.
I don't want to make assumptions about what data will be sent. In
general, carriage returns should not be preserved because I think that
"Hello World\r" is not what the client expects as the data of an event.

But it is also possible to send, for example, control commands to the
client, where "\r" is what the client is looking for. This is why I keep
a single carriage return.