Return 204 No Content if client should stop reconnecting
When CloseConnection is called, the client will reconnect after the
timespan specified by WaitRequest or the client's defaults. To prevent
further reconnection attempts, the server can send 204 No Content as per
HTML Standard 9.2.1.

A possible use case for this is when a real-time chat session is
terminated, but participants still want to join.

The best way to gracefully terminate a connection between client and
server is, in my opinion, to define an event that indicated termination.
This way, the client will be informed that no further events can be
expected without first having to reconnect.